added id field to User class
set logger middlerware in __init__

moved user exceptions to separate dir
