+++ backend/matcha/init_db.py
@@ -1,19 +1,19 @@
 import json
 import click
 
-from flask import current_app
+from flask import current_app as app
 from flask.cli import with_appcontext
 
-from matcha.handler.user import UserHandler
+from matcha.handlers.user import UserHandler
 
 def init_db_contents():
-    with current_app.open_resource('data.json') as f:
+    with app.open_resource('data.json') as f:
         data = json.loads(f.read().decode('utf-8'))
         handler = UserHandler()
 
         for user in data['users']:
             handler.register(user['username'], user['password'], user['email'], user['first_name'], user['last_name'])
-            click.echo(f'Created user {user["username"]}')
+            app.logger.info(f'Created user {user["username"]}')
 
             # TODO think of better way to avoid extra method call
             #user_id = get_user_id(engine, user['username'])
@@ -26,7 +26,7 @@ def init_db_contents():
 def init_db_contents_command():
     """Fill in data in fresh database"""
     init_db_contents()
-    click.echo('Initialized the database contents')
+    app.logger.info('Initialized the database contents')
 
 def run(app):
     app.cli.add_command(init_db_contents_command)
